+++ include/dgl/runtime/ndarray.h
@@ -6,7 +6,6 @@
 #ifndef DGL_RUNTIME_NDARRAY_H_
 #define DGL_RUNTIME_NDARRAY_H_
 
-#include <dgl/zerocopy_serializer.h>
 
 #include <atomic>
 #include <string>
@@ -162,12 +161,12 @@ class NDArray {
    * \param stream The input data stream
    * \return Whether load is successful
    */
-  inline bool Load(dmlc::Stream* stream);
+  bool Load(dmlc::Stream* stream);
   /*!
    * \brief Save NDArray to stream
    * \param stream The output data stream
    */
-  inline void Save(dmlc::Stream* stream) const;
+  void Save(dmlc::Stream* stream) const;
   /*!
    * \brief Create a NDArray that shares the data memory with the current one.
    * \param shape The shape of the new array.
@@ -241,6 +240,11 @@ class NDArray {
   template<typename T>
   std::vector<T> ToVector() const;
 
+#ifndef _WIN32
+  // static std::shared_ptr<SharedMemory> GetSharedMem(const NDArray& tensor);
+  std::shared_ptr<SharedMemory> GetSharedMem() const;
+#endif  // _WIN32
+
   /*!
    * \brief Function to copy data from one array to another.
    * \param from The source array.
@@ -268,11 +272,6 @@ class NDArray {
  */
 inline bool SaveDLTensor(dmlc::Stream* strm, const DLTensor* tensor);
 
-#ifndef _WIN32
-NDArray ZeroCopyLoadDLTensor(ZeroCopyStream* zc_strm);
-void ZeroCopySaveDLTensor(ZeroCopyStream* zc_strm, DLTensor* tensor,
-                          std::shared_ptr<SharedMemory> mem);
-#endif  // _WIN32
 
 /*!
  * \brief Reference counted Container object used to back NDArray.
@@ -466,78 +465,6 @@ inline bool SaveDLTensor(dmlc::Stream* strm,
   return true;
 }
 
-inline void NDArray::Save(dmlc::Stream* strm) const {
-  auto zc_strm = dynamic_cast<ZeroCopyStream*>(strm);
-  // Use ZeroCopySaveDLTensor when stream is a ZeroCopyStream
-  if (zc_strm) {
-#ifndef _WIN32
-    ZeroCopySaveDLTensor(zc_strm, const_cast<DLTensor*>(operator->()),
-                         this->data_->mem);
-#else
-    LOG(FATAL) << "ZeroCopyStream is not supported on windows";
-#endif  // _WIN32
-    return;
-  }
-  SaveDLTensor(strm, const_cast<DLTensor*>(operator->()));
-}
-
-inline bool NDArray::Load(dmlc::Stream* strm) {
-  auto zc_strm = dynamic_cast<ZeroCopyStream*>(strm);
-  // Use ZeroCopyLoadDLTensor when stream is a ZeroCopyStream
-  if (zc_strm) {
-#ifndef _WIN32
-    *this = ZeroCopyLoadDLTensor(zc_strm);
-    return true;
-#else
-    LOG(FATAL) << "ZeroCopyStream is not supported on windows";
-#endif  // _WIN32
-  }
-  uint64_t header, reserved;
-  CHECK(strm->Read(&header))
-      << "Invalid DLTensor file format";
-  CHECK(strm->Read(&reserved))
-      << "Invalid DLTensor file format";
-  CHECK(header == kDGLNDArrayMagic)
-      << "Invalid DLTensor file format";
-  DLContext ctx;
-  int ndim;
-  DLDataType dtype;
-  CHECK(strm->Read(&ctx))
-      << "Invalid DLTensor file format";
-  CHECK(strm->Read(&ndim))
-      << "Invalid DLTensor file format";
-  CHECK(strm->Read(&dtype))
-      << "Invalid DLTensor file format";
-  CHECK_EQ(ctx.device_type, kDLCPU)
-      << "Invalid DLTensor context: can only save as CPU tensor";
-  std::vector<int64_t> shape(ndim);
-  if (ndim != 0) {
-    CHECK(strm->ReadArray(&shape[0], ndim))
-        << "Invalid DLTensor file format";
-  }
-  NDArray ret = NDArray::Empty(shape, dtype, ctx);
-  int64_t num_elems = 1;
-  int elem_bytes = (ret->dtype.bits + 7) / 8;
-  for (int i = 0; i < ret->ndim; ++i) {
-    num_elems *= ret->shape[i];
-  }
-  int64_t data_byte_size;
-  CHECK(strm->Read(&data_byte_size))
-      << "Invalid DLTensor file format";
-  CHECK(data_byte_size == num_elems * elem_bytes)
-      << "Invalid DLTensor file format";
-  if (data_byte_size != 0)  {
-    // strm->Read will return the total number of elements successfully read.
-    // Therefore if data_byte_size is zero, the CHECK below would fail.
-    CHECK(strm->Read(ret->data, data_byte_size))
-        << "Invalid DLTensor file format";
-  }
-  if (!DMLC_IO_NO_ENDIAN_SWAP) {
-    dmlc::ByteSwap(ret->data, elem_bytes, num_elems);
-  }
-  *this = ret;
-  return true;
-}
 
 }  // namespace runtime
 }  // namespace dgl
